Repository Pattern

+++ TeachMe/resources/views/tickets/details.blade.php
@@ -29,18 +29,20 @@
                 @endforeach
             </p>
 
-            @if( ! currentUser()->hasVoted($ticket) )
-            {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
-                <button type="submit" class="btn btn-primary">
-                    <span class="glyphicon glyphicon-thumbs-up"></span> Votar
-                </button>
-            {!! Form::close() !!}
-            @else
-            {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
-                <button type="submit" class="btn btn-primary">
-                    <span class="glyphicon glyphicon-thumbs-down"></span> Quitar Voto
-                </button>
-            {!! Form::close() !!}
+            @if(Auth::check())
+                @if( ! currentUser()->hasVoted($ticket) )
+                {!! Form::open(['route' => ['votes.submit', $ticket->id], 'method' => 'POST']) !!}
+                    <button type="submit" class="btn btn-primary">
+                        <span class="glyphicon glyphicon-thumbs-up"></span> Votar
+                    </button>
+                {!! Form::close() !!}
+                @else
+                {!! Form::open(['route' => ['votes.destroy', $ticket->id], 'method' => 'DELETE']) !!}
+                    <button type="submit" class="btn btn-primary">
+                        <span class="glyphicon glyphicon-thumbs-down"></span> Quitar Voto
+                    </button>
+                {!! Form::close() !!}
+                @endif
             @endif
 
             <h3>Nuevo Comentario</h3>
